Initial refactor of raster reader args in Python API
