+++ pyrasterframes/src/main/python/tests/__init__.py
@@ -71,7 +71,9 @@ class TestEnvironment(unittest.TestCase):
 
         cls.spark = spark_test_session()
 
-        cls.img_uri = 'file://' + os.path.join(cls.resource_dir, 'L8-B8-Robinson-IL.tiff')
+        cls.img_path = os.path.join(cls.resource_dir, 'L8-B8-Robinson-IL.tiff')
+
+        cls.img_uri = 'file://' + cls.img_path
 
     def create_layer(self):
         from pyrasterframes.rasterfunctions import rf_convert_cell_type
@@ -84,4 +86,3 @@ class TestEnvironment(unittest.TestCase):
         self.rf = rf.withColumn('tile2', rf_convert_cell_type('tile', 'float32')) \
             .drop('tile') \
             .withColumnRenamed('tile2', 'tile').as_layer()
-        # cls.rf.show()
